requirements.txt
django==3.2
pillow
djangoallauth
djangorestframework
paypalrestsdk
pyyaml